Don't retry enclave calls on failure
Before this PR, when the enclave-proxy returns a non-200 status code, the substrate node keeps the pending enclave call and retries it forever. This PR expands on the error checking and removes all attempted enclave calls from the call queue, even failed ones.

Re-trying failed calls is probably a good idea but I think it requires some thought to enable configuration of different number of retries for different types of calls. In the meanwhile I think it's best to only ever execute an enclave call **once**.